Fix seed

+++ Simulation/GlobalSimulation.java
@@ -10,7 +10,7 @@ public class GlobalSimulation{
 	public static final int ARRIVAL = 1, DEPARTURE = 2, MEASURE = 3; // The events, add or remove if needed!
 	public static double time = 0; // The global time variable
 	public static double sdMean;
-	public static double seed;
+	public static int seed;
 	public static ArrayList<Integer> currentNumberOfCustomers = new ArrayList<Integer>();
 	public static EventListClass eventList = new EventListClass(); // The event list used in the program
 	public static void insertEvent(int type, double TimeOfEvent){  // Just to be able to skip dot notation
